Added both revisions into one folder
