+++ keyboards/lets_split/rev2/rev2.h
@@ -1,0 +1,25 @@
+#ifndef REV2_H
+#define REV2_H
+
+#include "quantum.h"
+
+void promicro_bootloader_jmp(bool program);
+
+#define KEYMAP( \
+	k00, k01, k02, k03, k04, k05, k45, k44, k43, k42, k41, k40, \
+	k10, k11, k12, k13, k14, k15, k55, k54, k53, k52, k51, k50, \
+	k20, k21, k22, k23, k24, k25, k65, k64, k63, k62, k61, k60, \
+	k30, k31, k32, k33, k34, k35, k75, k74, k73, k72, k71, k70 \
+	) \
+	{ \
+		{ k00, k01, k02, k03, k04, k05 }, \
+		{ k10, k11, k12, k13, k14, k15 }, \
+		{ k20, k21, k22, k23, k24, k25 }, \
+		{ k30, k31, k32, k33, k34, k35 }, \
+		{ k40, k41, k42, k43, k44, k45 }, \
+		{ k50, k51, k52, k53, k54, k55 }, \
+		{ k60, k61, k62, k63, k64, k65 }, \
+		{ k70, k71, k72, k73, k74, k75 } \
+	}
+
+#endif
